Menu Addition
Was able to implement a menu for the website, only problem is I can't see it, so that is something that needs to be worked on. Additionally, under the achievement breakdown, the selections are listing a bit more than I would like, so that'll have to be fixed as well.